pyproject.toml
[tool.poetry]
name = "social-music-app"
version = "0.1.0"
description = ""
authors = ["Maksim Vikhrov <maxvihr@yandex.ru>"]
readme = "README.md"
packages = [{include = "musicapp"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
djangorestframework = "^3.14.0"
mysql-connector-python = "8.0.32"
python-dotenv = "0.21.1"
requests = "2.28.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
